fix(atoms): texts -> text
