+++ src/components/atoms/new/Text/H1/index.stories.ts
@@ -1,13 +1,13 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { H2 } from '.'
+import { H1 } from '.'
 
 const defaultProps = {
   children: 'Heading HogeHoge Hoge',
 }
 
-const meta: Meta<typeof H2> = {
-  title: 'Atoms/Texts/H2',
-  component: H2,
+const meta: Meta<typeof H1> = {
+  title: 'Atoms/Text/H1',
+  component: H1,
   tags: ['text'],
   parameters: {
     layout: 'fullscreen',
@@ -15,7 +15,7 @@ const meta: Meta<typeof H2> = {
 }
 
 export default meta
-type Story = StoryObj<typeof H2>
+type Story = StoryObj<typeof H1>
 
 export const Default: Story = {}
 Default.args = defaultProps
