+++ src/components/atoms/new/Text/H2/index.module.scss
