+++ src/components/atoms/new/Text/P/index.tsx
